requirements.txt
GitPython
python-dotenv
langchain
openai
tiktoken
langchain_community